--- .vscode/settings.json
@@ -1,18 +1,0 @@
-{
-    "files.exclude": {
-        "**/.git": true,
-        "**/.svn": true,
-        "**/.hg": true,
-        "**/CVS": true,
-        "**/.DS_Store": true,
-        "**/Thumbs.db": true,
-        "**/.apivcs": true,
-        "**/.gitignore": true,
-        "**/.designer.json": true,
-        "**/.exchange.xml": true,
-        "**/.vscode": true,
-        "**/.mvn": true,
-        "**/.exchange_modules_tmp": true,
-        ".mule": true
-    }
-}
