+++ elementFactory/pageOne.ts
@@ -1,0 +1,7 @@
+const pageOne = {
+
+    pageTitle: 'h1[class="title"]',
+    industryHeading: 'h5[class="icon-box-heading"]'
+}
+export default{pageOne}
+
