+++ elementFactory/pageTwo.ts
@@ -1,0 +1,8 @@
+const pageTwo = {
+
+    homePageMenu: 'ul[id="menu-primary"] li a',
+    conatactUsPagelast: '(//div[@class="elementor-widget-container"]/div/h2)[5]',
+    contactUsText: 'div[class="elementor-widget-container"] h2'
+}
+export default {pageTwo}
+
