+++ utils/uiCommonMethodModule.ts
@@ -1,0 +1,10 @@
+
+import uiBase from '../common/uiBase';
+import apiBase from '../common/apiBase';
+
+const uiCommonUtils = {
+    uiBase,
+    apiBase,
+}
+export { uiCommonUtils }
+
