+++ utils/uiElementFactoryModule.ts
@@ -1,0 +1,16 @@
+
+import admin from '../elementFactory/pageOne';
+import home from '../elementFactory/pageTwo';
+
+// Define the pageFactoryUtils object
+const elementFactoryUtils = {
+    admin,
+    home,
+};
+
+// Export the pageFactoryUtils object
+export { elementFactoryUtils };
+
+
+
+
